refactor(bytes): move to single-export files
