+++ bytes/last_index_of_needle_test.ts
@@ -1,0 +1,42 @@
+// Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
+import { assertEquals } from "../testing/asserts.ts";
+import { lastIndexOfNeedle } from "./last_index_of_needle.ts";
+
+Deno.test("[bytes] lastIndexOfNeedle1", () => {
+  const i = lastIndexOfNeedle(
+    new Uint8Array([0, 1, 2, 0, 1, 2, 0, 1, 3]),
+    new Uint8Array([0, 1, 2]),
+  );
+  assertEquals(i, 3);
+});
+
+Deno.test("[bytes] lastIndexOfNeedle2", () => {
+  const i = lastIndexOfNeedle(
+    new Uint8Array([0, 1, 1]),
+    new Uint8Array([0, 1]),
+  );
+  assertEquals(i, 0);
+});
+
+Deno.test("[bytes] lastIndexOfNeedle3", () => {
+  const i = lastIndexOfNeedle(new Uint8Array(), new Uint8Array([0, 1]));
+  assertEquals(i, -1);
+});
+
+Deno.test("[bytes] lastIndexOfNeedle with start index", () => {
+  const i = lastIndexOfNeedle(
+    new Uint8Array([0, 1, 2, 0, 1, 2]),
+    new Uint8Array([0, 1]),
+    2,
+  );
+  assertEquals(i, 0);
+});
+
+Deno.test("[bytes] lastIndexOfNeedle with start index 2", () => {
+  const i = lastIndexOfNeedle(
+    new Uint8Array([0, 1, 2, 0, 1, 2]),
+    new Uint8Array([0, 1]),
+    -1,
+  );
+  assertEquals(i, -1);
+});
